Fix selection scrolling bug
Fix a scrolling bug where the new implementation of visual drawing
offset experiences an overflow and throws the program into an infinite
loop.

Add a different line to the CMakeLists file to change the building
of `compile_commands.json`, which was not working on my desktop.

Add more vim-style keybindings to the program's main loop. Horizontal
and vertical orientations (^hl$, gjkG).