add meminfo func, drawing func for main view
This commit contains the beginnings of formalizing system information in
the code!

Move memory info fields into dedicated struct and add one to the parent
sysinfo struct. Adds a function to update memory information that is
called in `procs_update`. Adds a function to draw system information in
the top of the main window.