add args struct to procinfo
Add an args struct to procinfo to hold complete information about the
arguments provided to the running process (this is also done in htop).
This data is collected from `/proc/[pid]/cmdline`.

This also adds in drawing support for the process detail window at the
bottom of the screen, as well as a function to print an array of strings
in `xutil.h`.