fix two bugs in proc_update
The first bug had something to do with the removal of terminated
processes in proc_update. As far as I can tell, the cursor would get
lost and cause the selected index and num_procs values to go below 0 and
wrap around, causing an infinte loop. I'm not really sure what fixed it
but it seems to be gone.

The second was in read_meminfo. Files weren't being closed so eventually
the system would run out (errno EMFILE, 24).